+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed player morphs not using correct alpha in some cases.
+
 ## [2.0.3] - 2023-08-26
 
 - Added "Give Magic Weapon", "Give Summon Weapon", and "Give Ranged Weapon" effects.